Replace several instances of ModifyVariable with ModifyArg

+++ src/main/java/dev/yurisuika/raised/mixin/client/gui/hud/ChatHudMixin.java
@@ -11,12 +11,12 @@ public class ChatHudMixin {
 
     @ModifyVariable(method = "mouseClicked", at = @At(value = "HEAD"), ordinal = 1, argsOnly = true)
     private double modifyMouseClick(double value) {
-        return value + (double)Raised.getChat();
+        return value + Raised.getChat();
     }
 
     @ModifyVariable(method = "toChatLineY", at = @At(value = "HEAD"), ordinal = 0, argsOnly = true)
     private double modifyChatTooltip(double value) {
-        return value + (double)Raised.getChat();
+        return value + Raised.getChat();
     }
 
     @ModifyVariable(method = "render", at = @At(value = "STORE"), ordinal = 6)
